Rewrite unit test as integration test
