+++ src/main/java/hudson/plugins/gradle/injection/BuildScanEnvironmentContributor.java
@@ -1,6 +1,5 @@
 package hudson.plugins.gradle.injection;
 
-import com.google.common.annotations.VisibleForTesting;
 import hudson.EnvVars;
 import hudson.Extension;
 import hudson.model.EnvironmentContributor;
@@ -9,7 +8,6 @@ import hudson.model.TaskListener;
 import hudson.util.Secret;
 
 import javax.annotation.Nonnull;
-import java.util.function.Supplier;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -20,20 +18,9 @@ public class BuildScanEnvironmentContributor extends EnvironmentContributor {
 
     public static final String GRADLE_ENTERPRISE_ACCESS_KEY = "GRADLE_ENTERPRISE_ACCESS_KEY";
 
-    private final Supplier<InjectionConfig> injectionConfigSupplier;
-
-    public BuildScanEnvironmentContributor() {
-        this(InjectionConfig.SUPPLIER);
-    }
-
-    @VisibleForTesting
-    BuildScanEnvironmentContributor(Supplier<InjectionConfig> injectionConfigSupplier) {
-        this.injectionConfigSupplier = injectionConfigSupplier;
-    }
-
     @Override
     public void buildEnvironmentFor(@Nonnull Run run, @Nonnull EnvVars envs, @Nonnull TaskListener listener) {
-        Secret secret = injectionConfigSupplier.get().getAccessKey();
+        Secret secret = InjectionConfig.get().getAccessKey();
         if (secret == null) {
             return;
         }
