+++ src/test/groovy/hudson/plugins/gradle/AbstractIntegrationTest.groovy
@@ -1,13 +1,20 @@
 package hudson.plugins.gradle
 
+import hudson.EnvVars
 import hudson.model.Cause
 import hudson.model.FreeStyleBuild
 import hudson.model.FreeStyleProject
+import hudson.model.Label
 import hudson.model.ParametersAction
 import hudson.model.ParametersDefinitionProperty
 import hudson.model.TextParameterDefinition
 import hudson.model.TextParameterValue
 import hudson.model.queue.QueueTaskFuture
+import hudson.plugins.gradle.injection.InjectionConfig
+import hudson.plugins.gradle.injection.NodeLabelItem
+import hudson.slaves.DumbSlave
+import hudson.slaves.EnvironmentVariablesNodeProperty
+import hudson.slaves.NodeProperty
 import org.jenkinsci.plugins.pipeline.maven.GlobalPipelineMavenConfig
 import org.jenkinsci.plugins.pipeline.maven.dao.PipelineMavenPluginDao
 import org.junit.rules.TestRule
@@ -15,7 +22,7 @@ import org.jvnet.hudson.test.FlagRule
 import org.jvnet.hudson.test.JenkinsRule
 import spock.lang.Specification
 
-class AbstractIntegrationTest extends Specification {
+abstract class AbstractIntegrationTest extends Specification {
 
     public final JenkinsRule j = new JenkinsRule()
 
@@ -29,6 +36,41 @@ class AbstractIntegrationTest extends Specification {
             useWorkspaceAsHome: true,
             switches          : '--no-daemon'
         ]
+    }
+
+    void restartSlave(DumbSlave slave) {
+        j.disconnectSlave(slave)
+        j.waitOnline(slave)
+    }
+
+    DumbSlave createSlave(String label) {
+        return j.createOnlineSlave(Label.get(label))
+    }
+
+    EnvVars withGlobalEnvVars(@DelegatesTo(EnvVars) Closure closure) {
+        NodeProperty nodeProperty = new EnvironmentVariablesNodeProperty()
+        EnvVars env = nodeProperty.getEnvVars()
+
+        closure.setDelegate(env)
+        closure.run()
+
+        j.jenkins.globalNodeProperties.clear()
+        j.jenkins.globalNodeProperties.add(nodeProperty)
+        env
+    }
+
+    InjectionConfig withInjectionConfig(@DelegatesTo(InjectionConfig) Closure closure) {
+        def config = InjectionConfig.get()
+
+        closure.setDelegate(config)
+        closure.run()
+
+        config.save()
+        config
+    }
+
+    static List<NodeLabelItem> labels(String... labels) {
+        return labels.collect { new NodeLabelItem(it) }
     }
 
     static QueueTaskFuture<FreeStyleBuild> triggerBuildWithParameter(FreeStyleProject p, String parameterName, String value) {
