+++ src/test/groovy/hudson/plugins/gradle/injection/BaseGradleInjectionIntegrationTest.groovy
@@ -1,11 +1,14 @@
 package hudson.plugins.gradle.injection
 
+import hudson.plugins.gradle.AbstractIntegrationTest
 import org.junit.Rule
 import org.junit.rules.RuleChain
 
-class BaseGradleInjectionIntegrationTest extends BaseInjectionIntegrationTest {
+/**
+ * Base class for tests that need a Jenkins instance and Gradle tool.
+ */
+abstract class BaseGradleInjectionIntegrationTest extends AbstractIntegrationTest {
 
     @Rule
     public final RuleChain rules = RuleChain.outerRule(noSpaceInTmpDirs).around(j).around(gradleInstallationRule)
-
 }
